requirements.txt
adium-theme-ubuntu==0.3.4
beautifulsoup4==4.5.1
configparser==3.5.0
Django==1.10.4
django-bootstrap3==8.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-extra-views==0.8.0
django-filter==1.0.1
django-guardian==1.4.6
django-modelcluster==2.0
django-tables2==1.3.0
django-taggit==0.18.3
django-treebeard==4.1.0
djangorestframework==3.5.3
ecdsa==0.13
enum34==1.1.6
flake8==3.3.0
jedi==0.10.0
mccabe==0.6.1
olefile==0.44
paramiko==1.16.0
pbr==1.10.0
pep8==1.7.0
Pillow==4.0.0
pycodestyle==2.3.1
pycrypto==2.6.1
pyflakes==1.5.0
pytz==2016.10
reportlab==3.3.0
requests==2.12.4
selenium==3.0.2
six==1.10.0
sqlparse==0.2.2
stevedore==1.19.1
Unidecode==0.4.19
unity-lens-photos==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wagtail==1.8
Willow==0.4
youtube-dl==2017.3.2
